requirements.txt
alembic==0.9.2
certifi==2017.4.17
chardet==3.0.4
click==6.7
defusedxml==0.5.0
factory-boy==2.8.1
Faker==0.7.16
Flask==0.12.2
Flask-Login==0.4.0
flask-marshmallow==0.8.0
Flask-Migrate==2.0.4
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.6
MarkupSafe==1.0
marshmallow==2.13.5
marshmallow-sqlalchemy==0.13.1
oauthlib==2.0.2
psycopg2==2.7.1
PyJWT==1.5.0
python-dateutil==2.6.0
python-editor==1.0.3
python3-openid==3.1.0
requests==2.17.3
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-flask==1.0.0
social-auth-app-flask-sqlalchemy==1.0.1
social-auth-core==1.3.0
social-auth-storage-sqlalchemy==1.1.0
SQLAlchemy==1.1.10
urllib3==1.21.1
Werkzeug==0.12.2
